# packages/server/src/cms/controllers/widget.controller.ts
import { Controller, Get, Query, Post, Put, Delete, Param } from '@nestjs/common';
import { Tags } from 'nest-swagger';
import { WidgetService } from './widget.service';
import { CreateWidgetDto, WidgetResponse, EditWidgetDto } from '../dto';
import { NullableParseIntPipe, ResultList } from '../../common';
import { Widget } from '../interfaces';
import { KeyValueDto } from './../../core/dto';

@Tags('cms')
@Controller('widget')
export class WidgetController {

    constructor(private readonly service: WidgetService) { }

    @Get('search')
    async search(
        @Query('keyword') keyword?: string,
        @Query('value') value?: string
    ): Promise<KeyValueDto[]> {
        return this.service.search(keyword, value);
    }

    @Post()
    async create(entry: CreateWidgetDto): Promise<WidgetResponse> {
        return this.service.create(entry);
    }

    @Put()
    async update(entry: EditWidgetDto): Promise<WidgetResponse> {
        return this.service.update(entry);
    }

    @Get('query')
    async query(
        @Query('keyword') keyword?: string,
        @Query('page', new NullableParseIntPipe()) page: number = 1,
        @Query('size', new NullableParseIntPipe()) size: number = 10,
    ): Promise<ResultList<Widget>> {
        return this.service.query(page, size, { keyword }, 'name');
    }

    @Delete(':id')
    async remove(@Param('id') id: string): Promise<boolean> {
        return this.service.remove(id);
    }

    @Get(':id')
    async findOne(@Param('id') id: string): Promise<Widget> {
        return this.service.findById(id);
    }
}