# packages/server/src/swagger/swagger.yaml
basePath: null
definitions:
...
    KeyValueDto:
        description: ""
        properties:
            label:
                type: string
                description: ""
            value:
                type: string
                description: ""
        type: object
        required:
            - label
            - value
    Query:
        description: ""
        properties:
            size:
                type: number
                format: double
                description: ""
            page:
                type: number
                format: double
                description: ""
        type: object
        required:
            - size
            - page
...
    CreateMenuReq:
        description: ""
        properties:
            name:
                type: string
                description: ""
            slug:
                type: string
                description: ""
            group:
                type: boolean
                description: ""
            link:
                type: string
                description: ""
            order:
                type: number
                format: double
                description: ""
            externalLink:
                type: string
                description: ""
            blank:
                type: boolean
                description: ""
            icon:
                type: string
                description: ""
            badge:
                type: string
                description: ""
            badgeDot:
                type: string
                description: ""
            badgeStatus:
                type: string
                description: ""
            enable:
                type: boolean
                description: ""
            expanded:
                type: boolean
                description: ""
            acl:
                type: string
                description: ""
            paths:
                type: array
                items:
                    type: object
                description: ""
            parent:
                type: string
                description: ""
            permissions:
                type: array
                items:
                    type: object
                description: ""
            isMenu:
                type: boolean
                description: ""
        type: object
        required:
            - name
            - slug
            - group
            - link
            - blank
            - enable
            - expanded
            - isMenu
    Menu:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            slug:
                type: string
                description: ""
            group:
                type: boolean
                description: ""
            link:
                type: string
                description: ""
            order:
                type: number
                format: double
                description: ""
            externalLink:
                type: string
                description: ""
            blank:
                type: boolean
                description: ""
            icon:
                type: string
                description: ""
            badge:
                type: string
                description: ""
            badgeDot:
                type: string
                description: ""
            badgeStatus:
                type: string
                description: ""
            enable:
                type: boolean
                description: ""
            expanded:
                type: boolean
                description: ""
            acl:
                type: string
                description: ""
            paths:
                type: array
                items:
                    type: object
                description: ""
            parent:
                type: object
                description: ""
            permissions:
                type: array
                items:
                    type: object
                description: ""
            isMenu:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - name
            - slug
            - group
            - link
            - order
            - externalLink
            - blank
            - icon
            - badge
            - badgeDot
            - badgeStatus
            - enable
            - expanded
            - acl
            - paths
            - parent
            - isMenu
    EditMenuReq:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            slug:
                type: string
                description: ""
            group:
                type: boolean
                description: ""
            link:
                type: string
                description: ""
            order:
                type: number
                format: double
                description: ""
            externalLink:
                type: string
                description: ""
            blank:
                type: boolean
                description: ""
            icon:
                type: string
                description: ""
            badge:
                type: string
                description: ""
            badgeDot:
                type: string
                description: ""
            badgeStatus:
                type: string
                description: ""
            enable:
                type: boolean
                description: ""
            expanded:
                type: boolean
                description: ""
            acl:
                type: string
                description: ""
            paths:
                type: array
                items:
                    type: object
                description: ""
            parent:
                type: string
                description: ""
            permissions:
                type: array
                items:
                    type: object
                description: ""
            isMenu:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - name
            - slug
            - group
            - link
            - blank
            - enable
            - expanded
            - isMenu
    ResultList$Menu:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/Menu'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
...
    Log:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            operator:
                type: string
                description: ""
            operatorIp:
                type: string
                description: ""
            operation:
                type: string
                description: ""
            comment:
                type: string
                description: ""
            createdAt:
                type: string
                format: date-time
                description: ""
        type: object
        required:
            - id
            - name
            - operator
            - operatorIp
            - operation
            - comment
            - createdAt
    ResultList$Log:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/Log'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
...
    CreateGroupReq:
        description: ""
        properties:
            outid:
                type: number
                format: double
                description: ""
            name:
                type: string
                description: ""
            icon:
                type: string
                description: ""
            parent:
                type: string
                description: ""
            paths:
                type: array
                items:
                    type: object
                description: ""
            director:
                type: string
                description: ""
            order:
                type: number
                format: double
                description: ""
            isRegion:
                type: boolean
                description: ""
            description:
                type: string
                description: ""
        type: object
        required:
            - name
            - order
    Group:
        description: ""
        properties:
            id:
                type: string
                description: ""
            outid:
                type: number
                format: double
                description: ""
            name:
                type: string
                description: ""
            icon:
                type: string
                description: ""
            parent:
                type: string
                description: ""
            paths:
                type: array
                items:
                    type: object
                description: ""
            director:
                type: string
                description: ""
            order:
                type: number
                format: double
                description: ""
            isRegion:
                type: boolean
                description: ""
            description:
                type: string
                description: ""
        type: object
        required:
            - id
            - outid
            - name
            - icon
            - parent
            - paths
            - director
            - order
            - isRegion
            - description
    EditGroupReq:
        description: ""
        properties:
            id:
                type: string
                description: ""
            outid:
                type: number
                format: double
                description: ""
            name:
                type: string
                description: ""
            icon:
                type: string
                description: ""
            parent:
                type: string
                description: ""
            paths:
                type: array
                items:
                    type: object
                description: ""
            director:
                type: string
                description: ""
            order:
                type: number
                format: double
                description: ""
            isRegion:
                type: boolean
                description: ""
            description:
                type: string
                description: ""
        type: object
        required:
            - id
            - name
            - order
    ResultList$Group:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/Group'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
...
    WidgetResponse:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            title:
                type: string
                description: ""
            params:
                type: object
                description: ""
            type:
                type: string
                description: ""
        type: object
        required:
            - id
            - name
            - title
            - params
            - type
    CreateWidgetDto:
        description: ""
        properties:
            name:
                type: string
                description: ""
            title:
                type: string
                description: ""
            params:
                type: object
                description: ""
            type:
                type: string
                description: ""
        type: object
        required:
            - name
            - title
            - params
            - type
    EditWidgetDto:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            title:
                type: string
                description: ""
            params:
                type: object
                description: ""
            type:
                type: string
                description: ""
        type: object
        required:
            - id
            - name
            - title
            - params
            - type
    Widget:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            title:
                type: string
                description: ""
            type:
                type: string
                description: ""
            params:
                type: object
                description: ""
        type: object
        required:
            - id
            - name
            - title
            - type
            - params
    ResultList$Widget:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/Widget'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
info:
    description: 项目标准接口
    license:
        name: MIT
    title: 'rest api interface'
    version: 0.0.1
paths:
...
    /menu/:
        post:
            operationId: MenusCreate
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/CreateMenuReq'
        put:
            operationId: MenusUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Menu'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/EditMenuReq'
    /menu/search:
        get:
            operationId: MenusSearch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/KeyValueDto'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: value
                    required: false
                    type: string
    /menu/query:
        get:
            operationId: MenusQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$Menu'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: isMenu
                    required: true
                    type: boolean
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
...
    '/menu/{id}':
        get:
            operationId: MenusFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Menu'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
    /log/query:
        get:
            operationId: LogsQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$Log'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
    '/log/{id}':
        get:
            operationId: LogsFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Log'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
...
    /group/:
        post:
            operationId: GroupsCreate
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/CreateGroupReq'
        put:
            operationId: GroupsUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Group'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/EditGroupReq'
    /group/search:
        get:
            operationId: GroupsSearch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/KeyValueDto'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: value
                    required: false
                    type: string
    /group/query:
        get:
            operationId: GroupsQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$Group'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
    '/group/{id}':
        get:
            operationId: GroupsFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Group'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
...
    /widget/search:
        get:
            operationId: WidgetSearch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/KeyValueDto'
            description: ""
            tags:
                - cms
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: value
                    required: false
                    type: string
    /widget/:
        post:
            operationId: WidgetCreate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/WidgetResponse'
            description: ""
            tags:
                - cms
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/CreateWidgetDto'
        put:
            operationId: WidgetUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/WidgetResponse'
            description: ""
            tags:
                - cms
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/EditWidgetDto'
    /widget/query:
        get:
            operationId: WidgetQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$Widget'
            description: ""
            tags:
                - cms
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
    '/widget/{id}':
        delete:
            operationId: WidgetRemove
            produces:
                - text/html
            responses:
                '200':
                    description: Ok
                    schema:
                        type: boolean
            description: ""
            tags:
                - cms
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
        get:
            operationId: WidgetFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Widget'
            description: ""
            tags:
                - cms
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
swagger: '2.0'
securityDefinitions: {}
produces:
    - application/json